fix: monster spawntime load and save

+++ source/monster_brush.cpp
@@ -82,7 +82,7 @@ void MonsterBrush::drawMonster(BaseMap* map, Tile* tile, void* parameter) {
 			});
 			if (it == tile->monsters.end()) {
 				const auto monster = newd Monster(monster_type);
-				monster->setSpawnMonsterTime(*(int*)parameter);
+				monster->setSpawnMonsterTime(*(uint16_t*)parameter);
 				tile->monsters.emplace_back(monster);
 			}
 		}
